+++ Source/Tasks/PostUserProfile.java
@@ -19,6 +19,8 @@ public class PostUserProfile extends NetworkTask {
         String userId = work.getStringData(1).substring(1);
         String targetUserId = work.getStringData(3).substring(1);
 
+        setSendUserIfNodeExist(targetUserId);
+
         if (!myProfile.equals(targetUserId))
             return this;