+++ Source/Tasks/SendChatFile.java
@@ -22,13 +22,12 @@ public class SendChatFile extends NetworkTask {
     void setFromString() {
         setProperties(Integer.parseInt(work.getStringData(0)), 10, "snd-cf", "rec-cf");
 
-        if (isOriginalTask())
-            return;
-
         String userId = work.getStringData(1).substring(1);
         String targetUserId = work.getStringData(2).substring(1);
 
-        if (!myProfile.equals(targetUserId))
+        setSendUserIfNodeExist(targetUserId);
+
+        if (isOriginalTask() || !myProfile.equals(targetUserId))
             return;
 
         skipSend = true;
@@ -54,9 +53,6 @@ public class SendChatFile extends NetworkTask {
 
         setProperties(timeout, 10, "snd-cf", "rec-cf");
 
-        if (isOriginalTask())
-            return;
-
         byte[] _userId = Util.getNextDataOnSize(data, 16);
         data = Util.clearByteArrayOnSize(data, 16);
         String userId = Util.convertByteArrayToHexString(_userId);
@@ -65,7 +61,9 @@ public class SendChatFile extends NetworkTask {
         data = Util.clearByteArrayOnSize(data, 16);
         String targetUserId = Util.convertByteArrayToHexString(_targetUserId);
 
-        if (!myProfile.equals(targetUserId))
+        setSendUserIfNodeExist(targetUserId);
+
+        if (isOriginalTask() || !myProfile.equals(targetUserId))
             return;
 
         skipSend = true;